requirements.txt
arxiv==2.1.0
google-generativeai
python-frontmatter
